🎨  Update performance project structure
