+++ performances/src/main.ts
@@ -1,10 +1,10 @@
 import puppeteer, { Page } from 'puppeteer'
-
 import { formatProfilingResults } from './format'
-import { startProfiling } from './profiling'
-import { trackNetwork } from './trackNetwork'
+import { startProfiling } from './profilers/startProfiling'
 import { ProfilingResults, ProfilingOptions } from './types'
 import { startProxy } from './proxy'
+import { runWikipediaScenario, wikipediaScenarioDescription } from './scenarios/runWikipediaScenario'
+import { runTwitterScenario, twitterScenarioDescription } from './scenarios/runTwitterScenario'
 
 main().catch((error) => {
   console.error(error)
@@ -34,33 +34,15 @@ Options:
     proxy,
     startRecording,
   }
+  const scenarios: Array<[string, (page: Page, takeMeasurements: () => Promise<void>) => Promise<void>]> = [
+    [wikipediaScenarioDescription, runWikipediaScenario],
+    [twitterScenarioDescription, runTwitterScenario],
+  ]
 
-  const wikipediaResults = await profileScenario(options, runWikipediaScenario)
-
-  console.log(`
-# Wikipedia
-
-Illustrates a mostly static site scenario.
-
-* Navigate on three Wikipedia articles
-* Do a search (with dynamic autocompletion) and go to the first result
-
-${formatProfilingResults(wikipediaResults)}`)
-
-  const twitterResults = await profileScenario(options, runTwitterScenario)
-
-  console.log(`
-# Twitter
-
-Illustrates a SPA scenario.
-
-* Navigate to the top trending topics
-* Click on the first trending topic
-* Click on Top, Latest, People, Photos and Videos tabs
-* Navigate to the Settings page
-* Click on a few checkboxes
-
-${formatProfilingResults(twitterResults)}`)
+  for (const [description, runScenario] of scenarios) {
+    const results = await profileScenario(options, runScenario)
+    console.log(`${description}\n\n${formatProfilingResults(results)}`)
+  }
 
   proxy.stop()
 }
@@ -89,64 +71,6 @@ async function profileScenario(
     await browser.close()
   }
   return result
-}
-
-async function runWikipediaScenario(page: Page, takeMeasurements: () => Promise<void>) {
-  await page.goto('https://en.wikipedia.org/wiki/Event_monitoring')
-  await takeMeasurements()
-
-  await page.goto('https://en.wikipedia.org/wiki/Datadog')
-  await takeMeasurements()
-
-  await page.type('[type="search"]', 'median', {
-    // large delay to trigger the autocomplete menu at each key press
-    delay: 400,
-  })
-  await Promise.all([page.waitForNavigation(), page.keyboard.press('Enter')])
-  await takeMeasurements()
-
-  await page.goto('https://en.wikipedia.org/wiki/Ubuntu')
-  await takeMeasurements()
-
-  await page.goto('about:blank')
-}
-
-async function runTwitterScenario(page: Page, takeMeasurements: () => Promise<void>) {
-  const { waitForNetworkIdle } = trackNetwork(page)
-  await page.goto('https://twitter.com/explore')
-  await waitForNetworkIdle()
-
-  // Even if the network is idle, sometimes links take a bit longer to render
-  await page.waitForSelector('[data-testid="trend"]')
-  await takeMeasurements()
-  await page.click('[data-testid="trend"]')
-  await waitForNetworkIdle()
-  await takeMeasurements()
-
-  // Click on all tabs
-  const tabs = await page.$$('[role="tab"]')
-  for (const tab of tabs) {
-    await tab.click()
-    await waitForNetworkIdle()
-    await takeMeasurements()
-  }
-
-  await page.click('[aria-label="Settings"]')
-  await waitForNetworkIdle()
-  await takeMeasurements()
-
-  // Scroll to the bottom of the page, because some checkboxes may be hidden below fixed banners
-  await page.evaluate(`scrollTo(0, 100000)`)
-
-  // Click on all checkboxes except the first one
-  const checkboxes = await page.$$('input[type="checkbox"]')
-  for (const checkbox of checkboxes.slice(1)) {
-    await checkbox.click()
-    await waitForNetworkIdle()
-    await takeMeasurements()
-  }
-
-  await page.goto('about:blank')
 }
 
 async function setupSDK(page: Page, options: ProfilingOptions) {
