+++ performances/src/profilers/startNetworkProfiling.ts
@@ -1,0 +1,37 @@
+import { CDPSession, Protocol } from 'puppeteer'
+import { ProfilingOptions } from 'src/types'
+import { isSdkBundleUrl } from './startProfiling'
+
+export async function startNetworkProfiling(options: ProfilingOptions, client: CDPSession) {
+  await client.send('Network.enable')
+  options.proxy.stats.reset()
+  // Twitter is using a service worker intercepting our requests
+  await client.send('Network.setBypassServiceWorker', { bypass: true })
+  let sdkDownload = 0
+
+  const sdkRequestIds = new Set<string>()
+
+  const requestListener = ({ request, requestId }: Protocol.Network.RequestWillBeSentEvent) => {
+    if (isSdkBundleUrl(options, request.url)) {
+      sdkRequestIds.add(requestId)
+    }
+  }
+
+  const loadingFinishedListener = ({ requestId, encodedDataLength }: Protocol.Network.LoadingFinishedEvent) => {
+    if (sdkRequestIds.has(requestId)) {
+      sdkDownload += encodedDataLength
+    }
+  }
+
+  client.on('Network.requestWillBeSent', requestListener)
+  client.on('Network.loadingFinished', loadingFinishedListener)
+  return () => {
+    client.off('Network.requestWillBeSent', requestListener)
+    client.off('Network.loadingFinishedListener', loadingFinishedListener)
+
+    return {
+      upload: options.proxy.stats.getStatsByHost(),
+      download: sdkDownload,
+    }
+  }
+}
