+++ performances/src/scenarios/runTwitterScenario.ts
@@ -1,0 +1,51 @@
+import { Page } from 'puppeteer'
+import { trackNetwork } from '../trackNetwork'
+
+export const twitterScenarioDescription = `# Twitter
+
+Illustrates a SPA scenario.
+
+* Navigate to the top trending topics
+* Click on the first trending topic
+* Click on Top, Latest, People, Photos and Videos tabs
+* Navigate to the Settings page
+* Click on a few checkboxes
+`
+
+export async function runTwitterScenario(page: Page, takeMeasurements: () => Promise<void>) {
+  const { waitForNetworkIdle } = trackNetwork(page)
+  await page.goto('https://twitter.com/explore')
+  await waitForNetworkIdle()
+
+  // Even if the network is idle, sometimes links take a bit longer to render
+  await page.waitForSelector('[data-testid="trend"]')
+  await takeMeasurements()
+  await page.click('[data-testid="trend"]')
+  await waitForNetworkIdle()
+  await takeMeasurements()
+
+  // Click on all tabs
+  const tabs = await page.$$('[role="tab"]')
+  for (const tab of tabs) {
+    await tab.click()
+    await waitForNetworkIdle()
+    await takeMeasurements()
+  }
+
+  await page.click('[aria-label="Settings"]')
+  await waitForNetworkIdle()
+  await takeMeasurements()
+
+  // Scroll to the bottom of the page, because some checkboxes may be hidden below fixed banners
+  await page.evaluate(`scrollTo(0, 100000)`)
+
+  // Click on all checkboxes except the first one
+  const checkboxes = await page.$$('input[type="checkbox"]')
+  for (const checkbox of checkboxes.slice(1)) {
+    await checkbox.click()
+    await waitForNetworkIdle()
+    await takeMeasurements()
+  }
+
+  await page.goto('about:blank')
+}
